# chainhash, wire, btcutil, main: Memory efficient txhash