@kcalvinalvin 2023-11-06
Rebased

For ibd and block verification, I think https://github.com/btcsuite/btcd/pull/2023 is going to be better but that one's still in the works.