@Roasbeef 2023-12-15
Will tag the new version, then make a new PR updating relevant `go.mod`s, before a final PR that swaps this in everywhere.